Upgraded ajax calls to be more reliable
Made the ajax calls more reliable and faster loading.

Before, there was a bug that would trigger a button on and off
then back on after only one button press.  This was because when
the button was manually pressed, it was in the middle if an ajax
update from the server, thus returning the wrong value.