Added global var for seeds.rb file
When running db:seed, the model methods were triggering actions
with the AEM which wouldn't be running when seeding the database.
This causes errors and the seeds.rb to crash.

With the global var, if it is set to true, the model methods ignore
the AEM actions.

This commit also sets the correct input ports for Mitch_IN and
Talkback_IN, as well as updates some of the jack error messages
to display the full name of the port when the port cant connect.