Update to Rails 6
Update the project from Rails 5 to Rails 6. This update introduces two
major changes to the app: Rails credentials, and the Webpacker gem.

Rails credentials
-----------------

Rails 5.2 introduced a replacement for Secrets: [Credentials].
Credentials provides a method for saving environment variables in an
encrypted format that is safe to keep in the repository. This feature
depends on a master key, which must be shared out-of-band.

Webpacker
---------

The [Webpacker gem] allows for easy integration of [webpack] into Rails
apps. This gem is designed to work with the asset pipeline, so existing
JavaScript and SASS should continue to work.

[Credentials]: https://github.com/rails/rails/pull/30067
[Webpacker gem]: https://github.com/rails/webpacker
[webpack]: https://webpack.js.org/