Added Cmd::Debug model
This model helps out with outputing to the terminal for debugging.
The method 'print' in this model automatically pulls the name of
the model and the name of the method that is running and
outputs that data along with a custom message that corrolates with
the method.

This also provides this project with a standardized way of outputting
verbose information to the terminal which will help with keeping
things organized.

For example:
'self.print "Jack is already running, stopping jack.."'

and the result in the terminal:
"  Cmd::Jack.start      > Jack is already running, stopping jack.."

This also keeps the title of the message to 20 chars to keep all the
output from these messages to the same starting location.  This helps
with readability and clenliness in the output.

This also indents the message by 2 spaces to differenciate between
the other output while rails is running.  It is a tad easier to spot
quickly this way.