Added Cmd::Jack Model
This model is responsible for controlling the Jack Audio interface
for the hipbox system.

This model can set the jack device, driver, sample rate, periods,
and buffer size, as well as starting and stopping the jack service.

It also provides a status method for checking if jack is already
running.

Modularizing this from the rest of the rails code should provide
a better level of upgradeability for this code in the future if
we decide to approach jack or the audio processing in a different
way.