Improve testing and development story
Add tools and libraries to the project that I think will make it easier
to develop and test our code.

Testing story
-------------

This project is now configured for automated testing with the RSpec test
framework. The RSpec installation has replaced the default test
framework (MiniTest).

A few gems for testing have been added:

- [`rspec-rails`]: provides RSpec/Rails compatibility
- [`action-cable-testing`]: utilities for testing ActionCable (more
  info below)
- [`capybara`]: utilities for acceptance testing
- [`factory_bot_rails`]: fixtures library
- [`shoulda-matchers`]: one-line methods for common tests
- [`webdrivers`]: for testing JavaScript

Rails 6 includes support for testing ActionCable, which has been enabled
for this project. More information about how to test ActionCable can be
found in the [action-cable-testing gem's README].

Development story
-----------------

In addition to the improvements to testing, this commit also includes
some improvements to the development process in general.

- [`bullet`]: helps identify unnecessary queries
- [`pry-byebug`and `pry-rails`]: a fantastic Ruby REPL
- [`rack-mini-profiler`]: performance analysis
- [`rubocop`], `rubocop-performance`, and `rubocop-rails`: enables us
  to check and enforce a consistent style

The RuboCop configuration included in this commit is based solely on my
own preferences, but we should evolve it going forward to suite both of
our preferences.

Personally, I like to run RuboCop from a pre-commit hook to make sure my
commits have a consistent style. This script can be found in
`bin/pre-commit`.

[`action-cable-testing`]: https://github.com/palkan/action-cable-testing
[`bullet`]: https://github.com/flyerhzm/bullet
[`capybara`]: https://github.com/teamcapybara/capybara
[`factory_bot_rails`]: https://github.com/thoughtbot/factory_bot_rails
[`pry-byebug` and `pry-rails`]: https://github.com/pry/pry
[`rack-mini-profiler`]: https://github.com/MiniProfiler/rack-mini-profiler
[`rspec-rails]: https://github.com/rspec/rspec-rails
[`rubocop`]: https://github.com/rubocop-hq/rubocop
[`shoulda-matchers`]: https://github.com/thoughtbot/shoulda-matchers
[`webdrivers`]: https://github.com/titusfortner/webdrivers