Modularized channel strips in channel_index view
The reason I did this was to break up the monolith of code in the
channel_index view.

There were a lot of shared components between the channel strips
and the master bus that were coded multiple times.  This breaks
down the components into partials and can be re-used by both
the master bus and the channel strips, thus reducing code and
complexity

This commit also removes all traces of JS inside the views.  Before,
I had script tags to load JS on page load which is now removed.  I
also had action tags such as 'onClick' on buttons and sliders which
have now also been removed in lieu of event listeners placed at
page load.  This cleaned up a large amount of useless HTML code.