Convert JS to ES6-style imports and exports
Now that the asset pipeline has been removed, some changes to JS loading
are needed to ensure that Webpack can find, bundle, and serve all the
necessary JavaScript.

This reorganization involves a few steps:

- Organizing JS code into "packs", which can then be included with the
  `javascript_pack_tag` method
- Converting all remaining JS functionality into ES6 modules
- Replacing inline script elements and HTML event attributes with
  regular JS code, which can be bundled and served

While all the JS is currently working, I think there are opportunities
to make improvements, specifically to the JS related to ActionCable. I
get the sense that we might be fighting against Rails, or maybe we're
under-utilizing the framework. Either way, we should keep an eye on
things.