Refactored all the JS
I really didn't want to do this big of a commit but all of this stuff
ties in together and there really isnt a good way to split this all
into individual commits with out things getting messy.. so here it is.

All the JS has been broken up from the original channel_index.js file.
This file was way too big, consisted of all orphaned functions, and
had a lot of redundant code.  The actioncable code was also inside
this file which should realistically be segregated in it's own file
in the javascripts/channels/ dir.

This commit breaks all that code into individual files pertaining
to the different channel strip components, ie. button, slider, readout.
This also adds namespaces to all the functions to keep things organized
and segregated the actioncable stuff into the javascripts/channels/
dir.

Method names have also been changed in the
app/channels/channels_channel.rb file.  These names now make a lot more
sense and coincide with the component names in the view all the way
through the code.  Before, I had different names for the same component
which, as you can imagine, created a huge mess.

This also complies with new ES6 code for import/export.