Prevent Turbolinks on specific links
On the links in the menubar to go back and forth between the
admin screen and mixer screen, turbolinks was preventing the actioncable
and JS from disconnecting and this was causing JS errors.  This patch
prevents turbolinks from working on these links.

Realistically, we might want to just get rid of turbolinks alltogether.