Refactored AEM code
This code was in desperate need of being broken up.  It consisted of too
many lines of code in a single file 'hipbox.cpp'.  This commit breaks up
that monolith into modules and separates them into separate categorical
directories.  This will provide us with better code management and allow
easier modifying and adding.