Added 'reset_data' function to data.h
This function deletes all routing objects and resets all vectors
to reset the AEM.  After this is done, the system will need to wait for
a json object from Rails in order to reload the AEM with the proper
routing objects, ie. ports, portGroups, users, etc.

This cleans up the rails_handler function.