Split up rails_handler function
Put a lot of the code in the rails_handler function back into the
structs for channels, mixers, inPorts, inPortGroups.. etc.

Now, you can create any of these objects by giving the object
a json block and it will construct everything for you.

This branch also standardizes the logging output by using 'PRINT' macro
so that everything looks pretty and neat.  To keep the standards here,
it is suggested to add the module.function> at the start of every PRINT
macro.  This lets the user know where in the code the log is coming
from.