Split up rails_handler into multiple OSC calls
OSC is supposed to be able to handle multiple types of incoming
data calls.  Before, I had all the calls going to one path, '/rails'.
This is not how OSC is supposed to be.

This commit starts to split the rails calls into a few groups.
Reset, mixer, and channel.  These groups are now split up into their
own functions.